backend/requirements.txt
deepagents==0.4.3
langchain==1.2.10
langgraph==1.0.9
openai==2.21.0
python-dotenv==1.2.1
fastapi==0.129.0
uvicorn==0.41.0
langchain-openai==1.1.10
python-frontmatter
langchain-mcp-adapters
fastmcp